% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuadratSampleGenerator.R
\name{QuadratSampleGenerator}
\alias{QuadratSampleGenerator}
\title{Quadrat Sample Generator}
\usage{
QuadratSampleGenerator(height, width, n_quadrats)
}
\arguments{
\item{height}{Integer for the height of the grid in the sampling area (in quadrats).}

\item{width}{Integer with the width of the grid in the sampling area (in quadrats).}

\item{n_quadrats}{Integer with the number quadrats to be sampled.}
}
\value{
A plot showing which quadrats to randomly sample in the specific grid.
}
\description{
Function randomly selects the desired number of quadrats to be sampled based
on the given dimensions of the gridded study site, and plots a map showing
these quadrats. Input should not be a vector (integer value only).
}
\examples{
quadrat_sample_generator(11, 10, 6)
quadrat_sample_generator (height = 5, width = 5, n_quadrats = 3)
}
